+++ tests/test_ne_reloc.c
@@ -8,9 +8,9 @@
  * verifying the patched segment bytes.
  *
  * Build with:
- *   gcc -std=c99 -Wall -Wextra -I../src ../src/ne_parser.c \
- *       ../src/ne_loader.c ../src/ne_reloc.c \
- *       test_ne_reloc.c -o test_ne_reloc
+ *   wcc -ml -za99 -wx -d2 -i=../src ../src/ne_parser.c ../src/ne_loader.c
+ *       ../src/ne_reloc.c test_ne_reloc.c
+ *   wlink system dos name test_ne_reloc.exe file test_ne_reloc.obj,ne_parser.obj,ne_loader.obj,ne_reloc.obj
  */
 
 #include "../src/ne_parser.h"
@@ -54,8 +54,8 @@ static int g_tests_failed = 0;
     do { \
         if ((a) != (b)) { \
             g_tests_failed++; \
-            printf("FAIL - expected %lld got %lld (line %d)\n", \
-                   (long long)(b), (long long)(a), __LINE__); \
+            printf("FAIL - expected %ld got %ld (line %d)\n", \
+                   (long)(b), (long)(a), __LINE__); \
             return; \
         } \
     } while (0)
@@ -64,8 +64,8 @@ static int g_tests_failed = 0;
     do { \
         if ((a) == (b)) { \
             g_tests_failed++; \
-            printf("FAIL - unexpected equal value %lld (line %d)\n", \
-                   (long long)(a), __LINE__); \
+            printf("FAIL - unexpected equal value %ld (line %d)\n", \
+                   (long)(a), __LINE__); \
             return; \
         } \
     } while (0)
